move plotting routines from `analysis.py` to `visualtization.py`, and include quadlet-tdc-dipole lattice in `beamlines.py`
