+++ examples/synthetic_6d/data/transforms-linear.yaml
