+++ phase_space_reconstruction/virtual/beamlines.py
@@ -1,7 +1,14 @@
 import torch
 import numpy as np
-from bmadx import PI, C_LIGHT
-from bmadx.bmad_torch.track_torch import TorchDrift, TorchQuadrupole, TorchCrabCavity, TorchSBend, TorchLattice
+from bmadx import PI
+from bmadx.bmad_torch.track_torch import (
+    TorchDrift,
+    TorchQuadrupole,
+    TorchCrabCavity,
+    TorchSBend,
+    TorchLattice
+    )
+
 
 def quad_drift(l_d = 1.0, l_q = 0.1, n_slices=5):
     '''Creates quad + drift lattice
@@ -105,7 +112,7 @@ def quad_tdc_bend(p0c, dipole_on=False):
     
     return lattice
 
-def quadlet_tdc_bend(p0c, dipole_on=False):
+def triplet_tdc_bend(p0c):
     # Design momentum
     p_design = p0c # eV/c
     
@@ -116,12 +123,9 @@ def quadlet_tdc_bend(p0c, dipole_on=False):
     
     # transverse deflecting cavity (TDC) parameters
     l_tdc = 0.23
-    #k_tdc = 2.00 # m^-1
     f_tdc = 1.3e9
-    #v_tdc = p_design * k_tdc * C_LIGHT / ( 2 * PI * f_tdc )
-    #v_tdc = 7.0e5
+    phi_tdc = 0.0 # zero-crossing phase
     v_tdc = 0.0 # scan parameter
-    phi_tdc = 0.0 # scan parameter (maybe?)
      
     # Bend parameters
     l_arc = 0.365 # arc length
@@ -132,67 +136,159 @@ def quadlet_tdc_bend(p0c, dipole_on=False):
     # Drifts with geometrical corrections: 
 
     # Drift from Quad to TDC
+    l_d3 = 0.7625 - l_q/2 - l_tdc/2
+
+    # Drift from TDC to Bend
+    l_d4 = 0.633 - l_tdc/2 - l_bend/2
+
+    # Drift from Bend to YAG 2 (bend on)
+    l_d5 = 0.895 - l_bend/2/np.cos(theta)
+    
+    # Elements:
+    q1 = TorchQuadrupole(
+        L = torch.tensor(l_q),
+        K1 = torch.tensor(k1),
+        NUM_STEPS = 5
+        )
+    
+    d1 = TorchDrift( L = torch.tensor(d_q) )
+
+    q2 = TorchQuadrupole(
+        L = torch.tensor(l_q),
+        K1 = torch.tensor(k1),
+        NUM_STEPS = 5
+        )
+    
+    d2 = TorchDrift( L = torch.tensor(d_q) )
+
+    q3 = TorchQuadrupole(
+        L = torch.tensor(l_q),
+        K1 = torch.tensor(k1),
+        NUM_STEPS = 5
+        )
+    
+    d3 = TorchDrift( L = torch.tensor(l_d3) )
+    
+    tdc = TorchCrabCavity(
+        L = torch.tensor(l_tdc),
+        VOLTAGE = torch.tensor(v_tdc),
+        RF_FREQUENCY = torch.tensor(f_tdc),
+        PHI0 = torch.tensor(phi_tdc),
+        TILT=torch.tensor(PI/2)
+        )
+    
+    d4 = TorchDrift( L = torch.tensor(l_d4) )
+
+    bend = TorchSBend(
+        L = torch.tensor(l_arc),
+        P0C = torch.tensor(p_design),
+        G = torch.tensor(g),
+        E1 = torch.tensor(theta/2),
+        E2 = torch.tensor(theta/2),
+        FRINGE_AT = "no_end"
+        )
+
+    d5 = TorchDrift( L = torch.tensor(l_d5) )
+    
+    lattice = TorchLattice(
+        [q1, d1, q2, d2, q3, d3, tdc, d4, bend, d5]
+        )
+    
+    return lattice
+
+def quadlet_tdc_bend(p0c, dipole_on = False):
+    # Design momentum
+    p_design = p0c # eV/c
+    
+    # Quadrupole parameters
+    l_q = 0.1
+    k1 = 0.0 
+    d_q = 0.1 #distance between quads in triplet
+    
+    # transverse deflecting cavity (TDC) parameters
+    l_tdc = 0.23
+    f_tdc = 1.3e9
+    phi_tdc = 0.0 # zero-crossing phase
+    v_tdc = 0.0 # scan parameter
+     
+    # Bend parameters
+    # fixed: 
+    l_bend = 0.365
+    # variable when on/off: 
+    if dipole_on:
+        theta = 20.0 * PI / 180.0
+        l_arc = l_bend * theta / (2 * np.sin(theta/2))
+        g = theta / l_arc 
+    if not dipole_on:
+        g = 2.22e-16 # machine epsilon to avoid numerical error
+        theta = 2*np.arcsin(l_bend*g/2)
+        l_arc = theta/g
+
+    # Drift from Quad to TDC
     l_d4 = 0.7625 - l_q/2 - l_tdc/2
 
     # Drift from TDC to Bend
     l_d5 = 0.633 - l_tdc/2 - l_bend/2
 
-    # Drift from Bend to YAG 2 (bend on)
-    l_d6_on = 0.895 - l_bend/2/np.cos(theta)
-
-    # Drift from Bend to straight ahead YAG (bend off)
-    l_d6_off = 0.895 - l_bend/2
+    # Drift from Bend to YAG 2 (corrected for dipole on/off)
+    l_d6 = 0.895 - l_bend/2/np.cos(theta)
     
     # Elements:
-    q1 = TorchQuadrupole(L = torch.tensor(l_q),
-                         K1 = torch.tensor(k1),
-                         NUM_STEPS = 5)
+    q1 = TorchQuadrupole(
+        L = torch.tensor(l_q),
+        K1 = torch.tensor(k1),
+        NUM_STEPS = 5
+        )
     
-    d1 = TorchDrift(L = torch.tensor(d_q))
+    d1 = TorchDrift( L = torch.tensor(d_q) )
 
-    q2 = TorchQuadrupole(L = torch.tensor(l_q),
-                         K1 = torch.tensor(k1),
-                         NUM_STEPS = 5)
+    q2 = TorchQuadrupole(
+        L = torch.tensor(l_q),
+        K1 = torch.tensor(k1),
+        NUM_STEPS = 5
+        )
     
-    d2 = TorchDrift(L = torch.tensor(d_q))
+    d2 = TorchDrift( L = torch.tensor(d_q) )
 
-    q3 = TorchQuadrupole(L = torch.tensor(l_q),
-                         K1 = torch.tensor(k1),
-                         NUM_STEPS = 5)
+    q3 = TorchQuadrupole(
+        L = torch.tensor(l_q),
+        K1 = torch.tensor(k1),
+        NUM_STEPS = 5
+        )
     
-    d3 = TorchDrift(L = torch.tensor(d_q))
+    d3 = TorchDrift( L = torch.tensor(d_q) )
 
-    q4 = TorchQuadrupole(L = torch.tensor(l_q),
-                         K1 = torch.tensor(k1),
-                         NUM_STEPS = 5)
+    q4 = TorchQuadrupole(
+        L = torch.tensor(l_q),
+        K1 = torch.tensor(k1),
+        NUM_STEPS = 5
+        )
     
-    d4 = TorchDrift(L = torch.tensor(l_d4))
+    d4 = TorchDrift( L = torch.tensor(l_d4) )
     
-    tdc = TorchCrabCavity(L = torch.tensor(l_tdc),
-                          VOLTAGE = torch.tensor(v_tdc),
-                          RF_FREQUENCY = torch.tensor(f_tdc),
-                          PHI0 = torch.tensor(phi_tdc),
-                          TILT=torch.tensor(PI/2)
-                         )
+    tdc = TorchCrabCavity(
+        L = torch.tensor(l_tdc),
+        VOLTAGE = torch.tensor(v_tdc),
+        RF_FREQUENCY = torch.tensor(f_tdc),
+        PHI0 = torch.tensor(phi_tdc),
+        TILT=torch.tensor(PI/2)
+        )
     
-    d5 = TorchDrift(L = torch.tensor(l_d5))
+    d5 = TorchDrift( L = torch.tensor(l_d5) )
 
-    if dipole_on:
-        bend = TorchSBend(L = torch.tensor(l_arc),
-                          P0C = torch.tensor(p_design),
-                          G = torch.tensor(g),
-                          E1 = torch.tensor(theta/2),
-                          E2 = torch.tensor(theta/2),
-                          FRINGE_AT = "no_end"
-                         )
-    
-        d6 = TorchDrift(L = torch.tensor(l_d6_on))
+    bend = TorchSBend(
+        L = torch.tensor(l_arc),
+        P0C = torch.tensor(p_design),
+        G = torch.tensor(g),
+        E1 = torch.tensor(theta/2),
+        E2 = torch.tensor(theta/2),
+        FRINGE_AT = "no_end"
+        )
 
-    else:
-        bend = TorchDrift(L = torch.tensor(l_bend))
-        d6 = TorchDrift(L = torch.tensor(l_d6_off))
-
+    d6 = TorchDrift( L = torch.tensor(l_d6) )
     
-    lattice = TorchLattice([q1, d1, q2, d2, q3, d3, q4, d4, tdc, d5, bend, d6])
+    lattice = TorchLattice(
+        [q1, d1, q2, d2, q3, d3, q4, d4, tdc, d5, bend, d6]
+        )
     
     return lattice
